Split oversized required-gate files
