+++ .pm/tasks/task_212396995cf3409eb34c8e9bec563ca3.execution.md
@@ -558,3 +558,12 @@ Example:
 - Expected Result: The GitHub `required-gate` doc-governance check no longer reports missing sections or newly added non-template project task rows for the unified-world seed/fragment docs.
 - Actual Result: `doc-governance-check: OK`; `git diff --check` passed.
 - Blocker / Next Action: Commit this evidence entry, push the PR branch, and continue normal PR CI/comment/merge watch.
+
+## 2026-06-21 01:38:00 CST / tpm
+- 完成内容: Repaired PR #547 required-gate Rust file-size failure.
+- 遗留事项: None for the reported file-size blocker.
+- Action: Split `ResolveCrisis` and `GrantMetaProgress` event derivation from `action_to_event_gameplay.rs` into the semantic helper module `action_to_event_gameplay_meta.rs`; split viewer auth bootstrap launcher tests from `oasis7_game_launcher_tests.rs` into `launcher_viewer_auth_bootstrap_tests.rs`. No `include!` or `part*` structural slice was introduced.
+- Validation Command: `./scripts/check-rust-file-size.sh`; `./scripts/cargo-dev.sh fmt --all --check`; `git diff --check`; `./scripts/cargo-dev.sh test -p oasis7 --bin oasis7_game_launcher -- --nocapture`; `./scripts/cargo-dev.sh test -p oasis7 --lib runtime_gameplay_action_claim_first_agent_registers_starter_agent -- --nocapture`; `./scripts/cargo-dev.sh test -p oasis7 --lib runtime_gameplay_action_claim_starter_oc_credits_liquid_oc_once -- --nocapture`; `./scripts/cargo-dev.sh test -p oasis7 --lib runtime_agent_chat_allows_zero_balance_after_starter_oc_claim -- --nocapture`.
+- Expected Result: The exact CI blocker `check-rust-file-size.sh` passes, both formerly oversized files are below 1200 lines, and the moved runtime/launcher behavior remains covered by focused tests.
+- Actual Result: `check-rust-file-size: OK` with oversized code/test files both 0; `action_to_event_gameplay.rs` is 1174 lines and `oasis7_game_launcher_tests.rs` is 1062 lines. Formatter, diff check, launcher bin tests (112 passed, 1 ignored), and all three focused runtime tests passed. Existing oasis7_node unused warnings remain unrelated.
+- Blocker / Next Action: Commit and push this CI fix, then continue PR checks/comment/merge watch.
