+++ crates/oasis7/src/runtime/world/event_processing/action_to_event_gameplay.rs
@@ -1147,132 +1147,27 @@ impl World {
                 crisis_id,
                 strategy,
                 success,
-            } => {
-                if !self.state.agents.contains_key(resolver_agent_id) {
-                    return Ok(WorldEventBody::Domain(DomainEvent::ActionRejected {
-                        action_id,
-                        reason: RejectReason::AgentNotFound {
-                            agent_id: resolver_agent_id.clone(),
-                        },
-                    }));
-                }
-                let crisis_id = crisis_id.trim();
-                if crisis_id.is_empty() {
-                    return Ok(WorldEventBody::Domain(DomainEvent::ActionRejected {
-                        action_id,
-                        reason: RejectReason::RuleDenied {
-                            notes: vec!["crisis_id cannot be empty".to_string()],
-                        },
-                    }));
-                }
-                let Some(crisis) = self.state.crises.get(crisis_id) else {
-                    return Ok(WorldEventBody::Domain(DomainEvent::ActionRejected {
-                        action_id,
-                        reason: RejectReason::RuleDenied {
-                            notes: vec![format!("crisis not found: {crisis_id}")],
-                        },
-                    }));
-                };
-                if crisis.status != CrisisStatus::Active {
-                    return Ok(WorldEventBody::Domain(DomainEvent::ActionRejected {
-                        action_id,
-                        reason: RejectReason::RuleDenied {
-                            notes: vec![format!(
-                                "crisis is not active and cannot be resolved: {}",
-                                crisis_id
-                            )],
-                        },
-                    }));
-                }
-                if self.state.time > crisis.expires_at {
-                    return Ok(WorldEventBody::Domain(DomainEvent::ActionRejected {
-                        action_id,
-                        reason: RejectReason::RuleDenied {
-                            notes: vec![format!(
-                                "crisis expired at {} and cannot be resolved: {}",
-                                crisis.expires_at, crisis_id
-                            )],
-                        },
-                    }));
-                }
-                let strategy = strategy.trim();
-                if strategy.is_empty() {
-                    return Ok(WorldEventBody::Domain(DomainEvent::ActionRejected {
-                        action_id,
-                        reason: RejectReason::RuleDenied {
-                            notes: vec!["crisis strategy cannot be empty".to_string()],
-                        },
-                    }));
-                }
-                let severity = crisis.severity.max(1);
-                let impact = if *success {
-                    i64::from(severity).saturating_mul(CRISIS_BASE_IMPACT_PER_SEVERITY)
-                } else {
-                    -i64::from(severity).saturating_mul(CRISIS_BASE_IMPACT_PER_SEVERITY)
-                };
-                Ok(WorldEventBody::Domain(DomainEvent::CrisisResolved {
-                    resolver_agent_id: resolver_agent_id.clone(),
-                    crisis_id: crisis_id.to_string(),
-                    strategy: strategy.to_string(),
-                    success: *success,
-                    impact,
-                }))
-            }
+            } => self.resolve_crisis_action_to_event(
+                action_id,
+                resolver_agent_id,
+                crisis_id,
+                strategy,
+                *success,
+            ),
             Action::GrantMetaProgress {
                 operator_agent_id,
                 target_agent_id,
                 track,
                 points,
                 achievement_id,
-            } => {
-                if !self.state.agents.contains_key(operator_agent_id) {
-                    return Ok(WorldEventBody::Domain(DomainEvent::ActionRejected {
-                        action_id,
-                        reason: RejectReason::AgentNotFound {
-                            agent_id: operator_agent_id.clone(),
-                        },
-                    }));
-                }
-                if !self.state.agents.contains_key(target_agent_id) {
-                    return Ok(WorldEventBody::Domain(DomainEvent::ActionRejected {
-                        action_id,
-                        reason: RejectReason::AgentNotFound {
-                            agent_id: target_agent_id.clone(),
-                        },
-                    }));
-                }
-                let track = track.trim();
-                if track.is_empty() {
-                    return Ok(WorldEventBody::Domain(DomainEvent::ActionRejected {
-                        action_id,
-                        reason: RejectReason::RuleDenied {
-                            notes: vec!["meta progression track cannot be empty".to_string()],
-                        },
-                    }));
-                }
-                if *points == 0 {
-                    return Ok(WorldEventBody::Domain(DomainEvent::ActionRejected {
-                        action_id,
-                        reason: RejectReason::InvalidAmount { amount: *points },
-                    }));
-                }
-                let normalized_achievement = achievement_id.as_ref().map(|value| value.trim());
-                if normalized_achievement.is_some_and(|value| value.is_empty()) {
-                    return Ok(WorldEventBody::Domain(DomainEvent::ActionRejected {
-                        action_id,
-                        reason: RejectReason::RuleDenied {
-                            notes: vec!["achievement_id cannot be empty".to_string()],
-                        },
-                    }));
-                }
-                Ok(WorldEventBody::Domain(DomainEvent::MetaProgressGranted {
-                    operator_agent_id: operator_agent_id.clone(),
-                    target_agent_id: target_agent_id.clone(),
-                    track: track.to_string(),
-                    points: *points,
-                    achievement_id: normalized_achievement.map(str::to_string),
-                }))
-            }
+            } => self.grant_meta_progress_action_to_event(
+                action_id,
+                operator_agent_id,
+                target_agent_id,
+                track,
+                *points,
+                achievement_id.as_deref(),
+            ),
             _ => unreachable!("action_to_event_gameplay received unsupported action variant"),
         }
     }
